+++ tests/test_justetf_interface.py
@@ -263,6 +263,9 @@ def test_justetf_holdings_countries_wrong_security_type() -> None:
         ticker.justetf_holdings_countries
 
 
+@pytest.mark.skipif(
+    skip_test, reason="Skipping in GH action as it reaches the limit of requests"
+)
 @pytest.mark.parametrize(
     "isin",
     [
@@ -281,6 +284,12 @@ def test_justetf_holdings_sectors(isin: str) -> None:
     assert etf_holdings.shape[0] >= 2
     assert etf_holdings.shape[1] == 1
 
+    for i in range(etf_holdings.shape[0]):
+        assert etf_holdings.iloc[i, 0] != ""
+
+    # check the second column is not empty
+    assert etf_holdings.columns[0] != ""
+
 
 @pytest.mark.skipif(
     skip_test, reason="Skipping in GH action as it reaches the limit of requests"
@@ -297,8 +306,8 @@ def test_justetf_holdings_sectors_wrong_security_type() -> None:
         ticker.justetf_holdings_sectors
 
 
-@pytest.mark.skipif(
-    skip_test, reason="Skipping in GH action as it reaches the limit of requests"
+@pytest.mark.skip(
+    reason="JustETF website structure changed, Selenium can no longer find the price element"
 )
 @pytest.mark.parametrize(
     "isin",